+++ webui/react/src/pages/Cluster/ClusterHistoricalUsage.tsx
@@ -5,8 +5,6 @@ import React, { useEffect, useState } from 'react';
 import Section from 'components/Section';
 import { parseUrl } from 'routes/utils';
 import { getResourceAllocationAggregated } from 'services/api';
-import { sumArrays } from 'utils/array';
-import { secondToHour } from 'utils/time';
 
 import css from './ClusterHistoricalUsage.module.scss';
 import ClusterHistoricalUsageChart from './ClusterHistoricalUsageChart';
@@ -14,6 +12,7 @@ import ClusterHistoricalUsageCsvModal from './ClusterHistoricalUsageCsvModal';
 import ClusterHistoricalUsageFilters, {
   ClusterHistoricalUsageFiltersInterface,
 } from './ClusterHistoricalUsageFilters';
+import { mapResourceAllocationApiToChartSeries, ResourceAllocationChartSeries } from './utils';
 
 export const DEFAULT_RANGE_DAY = 14;
 export const DEFAULT_RANGE_MONTH = 6;
@@ -25,65 +24,8 @@ export enum GroupBy {
   Month = 'month',
 }
 
-interface ChartSeries {
-  groupedBy: GroupBy,
-  hoursByAgentLabel: Record<string, number[]>,
-  hoursByExperimentLabel: Record<string, number[]>,
-  hoursByResourcePool: Record<string, number[]>,
-  hoursByUsername: Record<string, number[]>,
-  hoursTotal: Record<string, number[]>,
-  time: string[],
-}
-
-const mapToChartSeries = (labelByPeriod: Record<string, number>[]): Record<string, number[]> => {
-  // 1. convert [periodIndex: {label: seconds}, ...] to {label: {periodIndex: hours}, ...}
-  const periodByLabelIndexed: Record<string, Record<number, number>> = {};
-  labelByPeriod.forEach((period, periodIndex) => {
-    Object.keys(period).forEach(label => {
-      periodByLabelIndexed[label] = {
-        ...(periodByLabelIndexed[label] || {}),
-        [periodIndex]: secondToHour(period[label]),
-      };
-    });
-  });
-
-  // 2. convert {label: {periodIndex: hours}, ...} to {label: [hours, ...], ...}
-  const periodByLabelIndexedFlat: Record<string, number[]> = {};
-  Object.keys(periodByLabelIndexed).forEach(label => {
-    periodByLabelIndexedFlat[label] = [];
-    for (let i = 0; i < labelByPeriod.length; i++) {
-      periodByLabelIndexedFlat[label].push(periodByLabelIndexed[label][i] || 0);
-    }
-  });
-
-  // 3. find top 5 labels
-  const topLabels = Object.keys(periodByLabelIndexedFlat).map(label => {
-    const hours = periodByLabelIndexedFlat[label].reduce((acc, val) => acc + val, 0);
-    return [ label, hours ];
-  })
-    .sort((a, b) => ((b[1] as number) - (a[1] as number)))
-    .slice(0, 5)
-    .map(item => item[0]);
-
-  // 4. sum non-top labels hours into "other labels"
-  let ret = {};
-  let otherLabels: number[] = [];
-  Object.keys(periodByLabelIndexedFlat).forEach(label => {
-    if (topLabels.includes(label)) {
-      ret = { ...ret, [label]: periodByLabelIndexedFlat[label] };
-    } else {
-      otherLabels = sumArrays(otherLabels, periodByLabelIndexedFlat[label]);
-    }
-  });
-  if (otherLabels.length > 0) {
-    ret = { ...ret, ['other labels']: otherLabels };
-  }
-
-  return ret;
-};
-
 const ClusterHistoricalUsage: React.FC = () => {
-  const [ chartSeries, setChartSeries ] = useState<ChartSeries|null>(null);
+  const [ chartSeries, setChartSeries ] = useState<ResourceAllocationChartSeries|null>(null);
   const [ filters, setFilters ] = useState<ClusterHistoricalUsageFiltersInterface>({
     afterDate: dayjs().subtract(1 + DEFAULT_RANGE_DAY, 'day'),
     beforeDate: dayjs().subtract(1, 'day'),
@@ -186,17 +128,9 @@ const ClusterHistoricalUsage: React.FC = () => {
         startDate: filters.afterDate,
       });
 
-      const entries = res.resourceEntries;
-
-      setChartSeries({
-        groupedBy: filters.groupBy,
-        hoursByAgentLabel: mapToChartSeries(entries.map(item => item.byAgentLabel)),
-        hoursByExperimentLabel: mapToChartSeries(entries.map(item => item.byExperimentLabel)),
-        hoursByResourcePool: mapToChartSeries(entries.map(item => item.byResourcePool)),
-        hoursByUsername: mapToChartSeries(entries.map(item => item.byUsername)),
-        hoursTotal: { total: entries.map(item => secondToHour(item.seconds)) },
-        time: entries.map(item => item.periodStart),
-      });
+      setChartSeries(
+        mapResourceAllocationApiToChartSeries(res.resourceEntries, filters.groupBy),
+      );
     })();
   }, [ filters, isUrlParsed ]);
